FogBugz #128594: set gpio-dw and leds driver to be built in
Currently the gpio-dw driver and the leds driver are all set to be built
as modules.  Set them to be built-in.

Even though the LED driver is dependent on the GPIO drivers, it is fine
for the LED driver to be built in.  If the FPGA gpios and FPGA gpio LEDs
are added to the device tree, then the LED driver will defer loading until
all the gpio's that are listed as LEDs in the device tree are loaded.

Signed-off-by: Alan Tull <atull@altera.com>