FogBugz #154216: Fix common clock driver for SOCFPGA
For clocks with a fixed-divider, the "reg" parameter is not necessary.
The mpu_periph_clk does not have a gate, so change it to the type
"altr,socfpga-perip-clk".

Set the proper clock for the smp_twd timer in the device tree entry.
By doing this, revert the fix in kernel/smp_twd.c to match kernel.org.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>