FogBugz #163802: arch/nios2: Copy FDT from init memory to regular memory
If the supplied DTB is builtin into kernel, it is in __init memory and
will eventually be destroyed after bootup. But if drivers are loaded as
modules, they still need to access the DTB entries after the kernel has
started. Therefore, the DTB is moved to a new allocated area.
Thanks Jeroen Van den Keybus for the fix.

Signed-off-by: Ley Foon Tan <lftan@altera.com>