FogBugz #167813-1: Switch to use the s3c-hsotg USB driver for peripheral mode
The s3c-hsotg driver is compatible with the Designware dwc2 USB OTG core
that is on SOCFPGA. Switch over to use this driver as it is the mainline
kerne. The s3c-hsotg driver is much more robust and better supported than
the current version.

This patch removes the old dwc2 driver from the repo.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>