FogBugz #169570: TX TCP Checksum offload is broken for Jumbo frames
The Synopsys EMAC will checksum jumbo transmit frames only up to a
certain size correctly, since the controller will start to egress
an outbound packet when it starts reaching the maximum size of the
transmit fifo. In the Altera configuration case, the PBL, or
programmed burst length is set to 64, the transmit fifo size is 4K,
and the bus datawidth is 32-bits. The equation to determine the
correct maximum MTU is FIFOSIZE-((PBL+3)*(BUSWIDTH/8)) ==
4096-(67*4) = 3828. Then we need to account for possible VLAN tags,
inner and outer, and then transform than number into an Ethernet
MTU, which gives us 3828-8-14-4 (8 bytes for two VLANs, 14 bytes
for the standard Ethernet header size, and another 4 bytes for
CRC). That gives us 3802. So we'll make it a maximum jumbo frame
size of 3800.

Signed-off-by: Vince Bridgers <vbridger@altera.com>