FogBugz #175708: nios2: irq: Get mapped IRQ before calling the handler
The IRQ handler needs the Linux IRQ, not the hardware-specific IRQ.
Therefore, let's distinguish that clearly in the parameter name and also
call irq_find_mapping() to get the mapped IRQ before calling the handler.

This fixes a bootup issue, which caused handle_bad_irq() to be continously
called with IRQ = 0, which is the hwirq for the timer tick.

Tested on Altera's NEEK board with a custom FPGA design.

Signed-off-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>
Signed-off-by: Ley Foon Tan <lftan@altera.com>