FogBugz #176300: Add support for multiple clock parents
The periph and sdram PLL can have any of 3 parents. For the perip pll, the
parents can be either: osc1, osc2, or f2s_periph_ref_clk. For the sdram pll,
the parents can be: osc1, osc2, or f2s_sdram_ref_clk.

Update the clock driver to read the appropriate register for determine the
correct parent for the periph and sdram PLL.

Add the new clock entries into the socfpga DTSI file. Also remove the
"clock-frequency" property for the f2s_periph_ref_clk from the DTSI file, as
this property should be in the DTS file.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>