FogBugz #180994: Conditionally enable L2 cache ECC on startup
Due to a misunderstanding in the preloader/U-Boot handoff process
the L2 ECC was originally enabled in preloader. The fact that L2
cache is disabled upon entry into Linux means that the L2 cache
ECC can be enabled upon Linux startup. This patch enables L2 ECC
early in the startup before the L2 cache is enabled.

V2
- Leave aux_ctl initialization alone.
- Use of_iomap() function.

Signed-off-by: Thor Thayer <tthayer@altera.com>